vertexai_function_calling_demo/requirements.txt
fastapi==0.108.0
google-cloud-aiplatform==1.38.1
google-auth==2.26.2
itsdangerous==2.1.2
jinja2==3.1.3
markdown==3.5.2
types-Markdown==3.5.0.20240106
uvicorn[standard]==0.25.0
